'Keys' tests for 'stash' and 'enqueue'
